+++ frontend/src/routes/map/components/layer_style_menu/raster_option/DimensionSelector.svelte
@@ -5,7 +5,7 @@
 	import { onDestroy } from 'svelte';
 
 	import Accordion from '../../atoms/Accordion.svelte';
-	import RangeSlider from '../../atoms/RangeSlider.svelte';
+	import PlaybackControl from '../../atoms/PlaybackControl.svelte';
 	import TimeRuler from '../../atoms/TimeRuler.svelte';
 
 	import { ICONS } from '$lib/icons';
@@ -37,7 +37,6 @@
 	} from '$routes/map/utils/vector/dimension-runtime';
 	import { mapStore } from '$routes/stores/map';
 	import { isMobile } from '$routes/stores/ui';
-	import { slide } from 'svelte/transition';
 
 	type DimensionEnabledRasterEntry = MorivisRasterEntry<
 		RasterCategoricalStyle | RasterBaseMapStyle | RasterDemStyle | RasterTiffStyle | RasterCadStyle
@@ -94,7 +93,6 @@
 	let imageUpdateRequestId = 0;
 	let isUpdatingDimension = $state(false);
 	let playbackSpeed = $state(1200);
-	const playbackIntervalMs = $derived(2001 - playbackSpeed);
 
 	$effect(() => {
 		if (!dimension || dimensionState) return;
@@ -287,48 +285,10 @@
 		emblaMainCarousel.scrollPrev();
 	};
 
-	let isPlaying = $state(false);
-
-	let autoplayTimeout: ReturnType<typeof setTimeout> | null = null;
-
-	const clearAutoplayTimer = () => {
-		if (autoplayTimeout) {
-			clearTimeout(autoplayTimeout);
-			autoplayTimeout = null;
-		}
+	const handlePlaybackTick = async () => {
+		if (!emblaMainCarousel || isUpdatingDimension) return;
+		emblaMainCarousel.scrollNext();
 	};
-
-	const scheduleAutoplayTick = () => {
-		clearAutoplayTimer();
-		if (!isPlaying || !emblaMainCarousel) return;
-
-		autoplayTimeout = setTimeout(() => {
-			if (!isPlaying || !emblaMainCarousel) return;
-			if (!isUpdatingDimension) {
-				emblaMainCarousel.scrollNext();
-			}
-			scheduleAutoplayTick();
-		}, playbackIntervalMs);
-	};
-
-	const toggleAutoplay = () => {
-		if (!emblaMainCarousel) return;
-
-		if (isPlaying) {
-			clearAutoplayTimer();
-			isPlaying = false;
-		} else {
-			isPlaying = true;
-			scheduleAutoplayTick();
-		}
-	};
-
-	$effect(() => {
-		const autoplayIntervalMs = playbackIntervalMs;
-		if (!autoplayIntervalMs) return;
-		if (!isPlaying) return;
-		scheduleAutoplayTick();
-	});
 
 	// ホイールイベント用の変数
 	let carouselElement: HTMLElement | undefined = $state();
@@ -366,7 +326,6 @@
 	};
 
 	onDestroy(() => {
-		clearAutoplayTimer();
 		if (carouselElement) {
 			carouselElement.removeEventListener('wheel', handleWheel);
 		}
@@ -469,36 +428,13 @@
 				</div>
 			</div>
 			{#if dimension.type === 'time'}
-				<div class="flex items-center justify-center gap-2 pt-3">
-					<button
-						onclick={toggleAutoplay}
-						class="bg-sub flex w-[200px] cursor-pointer items-center justify-center gap-1 rounded-full p-1 text-sm text-white select-none hover:bg-white/10"
-						aria-label={isPlaying ? '停止' : '再生'}
-					>
-						{#if isPlaying}
-							<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24">
-								<path fill="currentColor" d="M6 19h4V5H6v14zm8-14v14h4V5h-4z" />
-							</svg>
-							停止
-						{:else}
-							<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24">
-								<path fill="currentColor" d="M8 5v14l11-7z" />
-							</svg>
-							再生
-						{/if}
-					</button>
+				<div class="pt-3">
+					<PlaybackControl
+						disabled={isUpdatingDimension}
+						bind:playbackSpeed
+						onTick={handlePlaybackTick}
+					/>
 				</div>
-				{#if isPlaying}
-					<div transition:slide class="pt-2">
-						<RangeSlider
-							label={`再生速度 (${Math.round((1000 / playbackIntervalMs) * 10) / 10} コマ/秒)`}
-							bind:value={playbackSpeed}
-							min={1}
-							max={2000}
-							step={1}
-						/>
-					</div>
-				{/if}
 			{/if}
 		{/if}
 	</Accordion>
